requirements.txt
chromadb==0.3.21
llama-cpp-python==0.1.48
